# Conflicts:
#	code/Ragl/graph_region.h
#	code/Ratl/map_vs.h
#	code/cgame/FxScheduler.cpp
#	code/cgame/FxTemplate.cpp
#	code/cgame/cg_event.cpp
#	code/cgame/cg_local.h
#	code/cgame/cg_main.cpp
#	code/cgame/cg_players.cpp
#	code/cgame/cg_view.cpp
#	code/client/client.h
#	code/client/snd_ambient.cpp
#	code/client/snd_dma.cpp
#	code/game/AI_Default.cpp
#	code/game/AI_HazardTrooper.cpp
#	code/game/NPC.cpp
#	code/game/NPC_combat.cpp
#	code/game/NPC_reactions.cpp
#	code/game/NPC_spawn.cpp
#	code/game/NPC_utils.cpp
#	code/game/Q3_Interface.cpp
#	code/game/b_local.h
#	code/game/bg_pmove.cpp
#	code/game/g_active.cpp
#	code/game/g_combat.cpp
#	code/game/g_itemLoad.cpp
#	code/game/g_local.h
#	code/game/g_main.cpp
#	code/game/g_navigator.cpp
#	code/game/g_ref.cpp
#	code/game/g_shared.h
#	code/game/g_svcmds.cpp
#	code/icarus/BlockStream.cpp
#	code/icarus/Sequence.cpp
#	code/icarus/Sequencer.cpp
#	code/icarus/TaskManager.cpp
#	code/qcommon/cm_local.h
#	code/qcommon/hstring.h
#	code/qcommon/strip.cpp
#	code/rd-vanilla/G2_API.cpp
#	code/rd-vanilla/G2_misc.cpp
#	code/rd-vanilla/tr_ghoul2.cpp
#	codeJK2/cgame/FxScheduler.cpp
#	codeJK2/cgame/FxScheduler.h
#	codeJK2/cgame/FxSystem.cpp
#	codeJK2/cgame/FxSystem.h
#	codeJK2/cgame/FxTemplate.cpp
#	codeJK2/cgame/cg_credits.cpp
#	codeJK2/cgame/cg_event.cpp
#	codeJK2/cgame/cg_main.cpp
#	codeJK2/cgame/cg_players.cpp
#	codeJK2/game/AI_Default.cpp
#	codeJK2/game/AI_Jedi.cpp
#	codeJK2/game/CMakeLists.txt
#	codeJK2/game/NPC.cpp
#	codeJK2/game/NPC_combat.cpp
#	codeJK2/game/NPC_reactions.cpp
#	codeJK2/game/NPC_spawn.cpp
#	codeJK2/game/NPC_stats.cpp
#	codeJK2/game/NPC_utils.cpp
#	codeJK2/game/Q3_Interface.cpp
#	codeJK2/game/Q3_Registers.h
#	codeJK2/game/b_local.h
#	codeJK2/game/bg_pmove.cpp
#	codeJK2/game/g_ICARUS.cpp
#	codeJK2/game/g_active.cpp
#	codeJK2/game/g_combat.cpp
#	codeJK2/game/g_local.h
#	codeJK2/game/g_navigator.cpp
#	codeJK2/game/g_navigator.h
#	codeJK2/game/g_ref.cpp
#	codeJK2/game/g_savegame.cpp
#	codeJK2/game/g_shared.h
#	codeJK2/game/g_spawn.cpp
#	codeJK2/game/g_weapon.cpp
#	codeJK2/game/genericparser2.cpp
#	codeJK2/game/genericparser2.h
#	codeJK2/game/teams.h
#	codeJK2/icarus/BlockStream.cpp
#	codeJK2/icarus/Interpreter.cpp
#	codeJK2/icarus/Sequence.cpp
#	codeJK2/icarus/Sequencer.cpp
#	codeJK2/icarus/TaskManager.cpp
#	codeJK2/icarus/Tokenizer.cpp
#	codeJK2/icarus/blockstream.h
#	codeJK2/icarus/instance.h
#	codeJK2/icarus/interpreter.h
#	codeJK2/icarus/sequence.h
#	codeJK2/icarus/sequencer.h
#	codeJK2/icarus/taskmanager.h
#	codeJK2/icarus/tokenizer.h
#	documentation/developer/save games.md

+++ codemp/qcommon/cm_local.h
@@ -45,10 +45,10 @@ typedef struct cLeaf_s {
 	int			cluster;
 	int			area;
 
-	int			firstLeafBrush;
+	intptr_t	firstLeafBrush;
 	int			numLeafBrushes;
 
-	int			firstLeafSurface;
+	intptr_t	firstLeafSurface;
 	int			numLeafSurfaces;
 } cLeaf_t;
